style(contracts): forge fmt the generated Verifier.sol
The Foundry CI formatting gate (forge fmt --check) failed on the P3.2 push: the
raw snarkjs-generated Verifier.sol didn't conform to the repo's forge fmt style
(blank-line + indentation rules around inline assembly).

forge fmt is whitespace-only (no semantic change); verified:
- forge build → "Compiler run successful!" (Solc 0.8.20)
- forge fmt --check → clean

Keeps the formatting gate green for the rest of Phase 3.

+++ contracts/src/Verifier.sol
@@ -22,17 +22,17 @@ pragma solidity >=0.7.0 <0.9.0;
 
 contract Groth16Verifier {
     // Scalar field size
-    uint256 constant r    = 21888242871839275222246405745257275088548364400416034343698204186575808495617;
+    uint256 constant r = 21888242871839275222246405745257275088548364400416034343698204186575808495617;
     // Base field size
-    uint256 constant q   = 21888242871839275222246405745257275088696311157297823662689037894645226208583;
+    uint256 constant q = 21888242871839275222246405745257275088696311157297823662689037894645226208583;
 
     // Verification Key data
-    uint256 constant alphax  = 4046810554511096347223471314503653405319010211549719891067341276410419554538;
-    uint256 constant alphay  = 20972216556230368226129142191999290211441534694053991095223717295099065119396;
-    uint256 constant betax1  = 1162525274628084786254208382708110717945079403327240881866527625099479139524;
-    uint256 constant betax2  = 17006775537206652401980775245876252329592576393762318603098098078744078304278;
-    uint256 constant betay1  = 728736684120689475367658410355586091946600404348752493917377386977690379436;
-    uint256 constant betay2  = 11247183439082070242085762446097360082332904929189577386647035313668925265922;
+    uint256 constant alphax = 4046810554511096347223471314503653405319010211549719891067341276410419554538;
+    uint256 constant alphay = 20972216556230368226129142191999290211441534694053991095223717295099065119396;
+    uint256 constant betax1 = 1162525274628084786254208382708110717945079403327240881866527625099479139524;
+    uint256 constant betax2 = 17006775537206652401980775245876252329592576393762318603098098078744078304278;
+    uint256 constant betay1 = 728736684120689475367658410355586091946600404348752493917377386977690379436;
+    uint256 constant betay2 = 11247183439082070242085762446097360082332904929189577386647035313668925265922;
     uint256 constant gammax1 = 11559732032986387107991004021392285783925812861821192530917403151452391805634;
     uint256 constant gammax2 = 10857046999023057135944570762232829481370756359578518086990519993285655852781;
     uint256 constant gammay1 = 4082367875863433681332203403145435568316851327593401208105741076214120093531;
@@ -42,27 +42,30 @@ contract Groth16Verifier {
     uint256 constant deltay1 = 17493471943993725873570604148170470014090613528400949801396631377352624204699;
     uint256 constant deltay2 = 21556488447235544264854028591039827591893642708016503354644241971213596748373;
 
-    
     uint256 constant IC0x = 6946138913752544930832292946937867915765399931984519108737293770426867800473;
     uint256 constant IC0y = 18242390803140851463172567909700208051523547026199479718264228402093812427578;
-    
+
     uint256 constant IC1x = 3014839076773597899756147928084567102718424733265059754823117498005967875607;
     uint256 constant IC1y = 9733134220805379567993797237904875121523973816568277677019496727000054922875;
-    
+
     uint256 constant IC2x = 13358383084798687547415321492146287938956151156176087245554772425975630395894;
     uint256 constant IC2y = 5954534970797791919267528090954102482910636357629528202472653299681733434185;
-    
+
     uint256 constant IC3x = 9850703961214285540823921855687918184934428552930711013401086970681205309423;
     uint256 constant IC3y = 13497241013209238223651503236936767620888972829910428902640938008930643099230;
-    
- 
+
     // Memory data
     uint16 constant pVk = 0;
     uint16 constant pPairing = 128;
 
     uint16 constant pLastMem = 896;
 
-    function verifyProof(uint[2] calldata _pA, uint[2][2] calldata _pB, uint[2] calldata _pC, uint[3] calldata _pubSignals) public view returns (bool) {
+    function verifyProof(
+        uint256[2] calldata _pA,
+        uint256[2][2] calldata _pB,
+        uint256[2] calldata _pC,
+        uint256[3] calldata _pubSignals
+    ) public view returns (bool) {
         assembly {
             function checkField(v) {
                 if iszero(lt(v, r)) {
@@ -70,7 +73,7 @@ contract Groth16Verifier {
                     return(0, 0x20)
                 }
             }
-            
+
             // G1 function to multiply a G1 value(x,y) to value in an address
             function g1_mulAccC(pR, x, y, s) {
                 let success
@@ -105,13 +108,12 @@ contract Groth16Verifier {
                 mstore(add(_pVk, 32), IC0y)
 
                 // Compute the linear combination vk_x
-                
+
                 g1_mulAccC(_pVk, IC1x, IC1y, calldataload(add(pubSignals, 0)))
-                
+
                 g1_mulAccC(_pVk, IC2x, IC2y, calldataload(add(pubSignals, 32)))
-                
+
                 g1_mulAccC(_pVk, IC3x, IC3y, calldataload(add(pubSignals, 64)))
-                
 
                 // -A
                 mstore(_pPairing, calldataload(pA))
@@ -137,7 +139,6 @@ contract Groth16Verifier {
                 mstore(add(_pPairing, 384), mload(add(pMem, pVk)))
                 mstore(add(_pPairing, 416), mload(add(pMem, add(pVk, 32))))
 
-
                 // gamma2
                 mstore(add(_pPairing, 448), gammax1)
                 mstore(add(_pPairing, 480), gammax2)
@@ -154,7 +155,6 @@ contract Groth16Verifier {
                 mstore(add(_pPairing, 704), deltay1)
                 mstore(add(_pPairing, 736), deltay2)
 
-
                 let success := staticcall(sub(gas(), 2000), 8, _pPairing, 768, _pPairing, 0x20)
 
                 isOk := and(success, mload(_pPairing))
@@ -164,19 +164,18 @@ contract Groth16Verifier {
             mstore(0x40, add(pMem, pLastMem))
 
             // Validate that all evaluations ∈ F
-            
+
             checkField(calldataload(add(_pubSignals, 0)))
-            
+
             checkField(calldataload(add(_pubSignals, 32)))
-            
+
             checkField(calldataload(add(_pubSignals, 64)))
-            
 
             // Validate all evaluations
             let isValid := checkPairing(_pA, _pB, _pC, _pubSignals, pMem)
 
             mstore(0, isValid)
-             return(0, 0x20)
-         }
-     }
- }
+            return(0, 0x20)
+        }
+    }
+}
